v0.4.177: split AssignmentRoutes into five RouteCollections
Phase 2 of the audit-driven refactor.

The old `struct AssignmentRoutes` extended over 17 `+*.swift` files and
~6.5 KLOC of handlers from five conceptually independent surfaces.
Swift type-checks every extension as part of the parent type, so every
edit to any of the 17 files forced revalidation of the whole struct.
Splitting into five smaller `RouteCollection`s gives the compiler five
smaller type-check scopes and matches the actual ownership of these
URLs on the instructor side of the app.

* `InstructorDashboardRoutes` — dashboard list, status, validate page,
  grade CSV, per-assignment submission drilldown, BrightSpace sync.
  Renamed from `AssignmentRoutes` (same files).
* `DraftAssignmentRoutes` — draft authoring (new-assignment page,
  draft suite / family / check / script / suite-section CRUD, save,
  publish).
* `PublishedAssignmentRoutes` — published-assignment editing
  (edit/save, file downloads, script CRUD, unified suite editor,
  suite-section CRUD, global variables, pattern families, notebook
  checks).  Also hosts the two `/instructor`-scope utilities used by
  both pages: `script-templates` and `scan-notebook`.
* `StudentCourseRoutes` — per-course, per-student submission views.
* `CourseAdminRoutes` — course section CRUD + roster management.

Three small support changes:

* Four nested DTOs on the old type (`SuitePayload`, `SuiteItemDTO`,
  `ScriptDTO`, `TestSuiteSectionDTO`) lifted to top-level
  `SuitePayloadDTOs.swift` so the draft and published collections can
  share them.
* `preferredResultsBySubmissionID` promoted from a method to a free
  function (shared by `InstructorDashboardRoutes` and `StudentCourseRoutes`).
* `draftSolutionNotebook` moved from `+Editor.swift` (now Published)
  to `+Draft.swift` to land with the rest of `DraftAssignmentRoutes`.

The 17 `AssignmentRoutes+*.swift` files keep their names for `git
blame` continuity; a follow-up pass will rename them to match their
new parent.

No behaviour change.  1273/1273 tests pass; 0 SwiftLint violations.
