+++ Sources/APIServer/Routes/Web/AssignmentRoutes+Checks.swift
@@ -12,7 +12,7 @@ import Fluent
 import Foundation
 import Vapor
 
-extension AssignmentRoutes {
+extension PublishedAssignmentRoutes {
 
     // MARK: - GET /instructor/:assignmentID/checks
     //
