+++ Sources/APIServer/routes.swift
@@ -48,7 +48,11 @@ func routes(_ app: Application) throws {
     // MARK: - Instructor or admin only
 
     let instructor = app.grouped(sessionAuth, RoleMiddleware(required: .instructor), csrf)
-    try instructor.register(collection: AssignmentRoutes())
+    try instructor.register(collection: InstructorDashboardRoutes())
+    try instructor.register(collection: DraftAssignmentRoutes())
+    try instructor.register(collection: PublishedAssignmentRoutes())
+    try instructor.register(collection: CourseAdminRoutes())
+    try instructor.register(collection: StudentCourseRoutes())
     try instructor.register(collection: MarmosetImportRoutes())
     // Worker job polling is instructor-tier: only the server operator runs workers.
     try instructor.register(collection: SubmissionRoutes())
